Redmine #7075 Store absolute time with process information
Processes in CFEngine are identified by their OS process identifier (pid)
and their start time.

Up to now, the start time was relative to the boot time, so it was not
unique across system restarts.

Using an absolute start time assures that the combination of pid and
start time is unique. This fixes bug #7075.

Signed-off-by: Stefan Weil <sw@weilnetz.de>